sparse: 0.5.0 -> 0.6.3 and fixes
Added sqlite and gtk+3 that are now wanted by sparse.

Sparse expects to find system includes at /usr/include redirect this
to glibc.dev.

Sparse uses gcc -print-file-name= extensively to find gcc includes, due
to split in nixos this direct to gcc-lib, which does not include the headers
so redirect to GCC_BASE as defined in derivation. There might be a better way
to do this, but I did not immediately find one.

Define PREFIX as Make parameter, as old sed expression was broken.

Add a trivial test for cgcc.

Slightly irrelevant ref.
https://bugzilla.redhat.com/show_bug.cgi?id=1114755