tar2ext4: init at 0.8.20
This is a single program that's part of a big Microsoft project.
Everything else looks pretty specialised, and some of it only targets
Windows, so I didn't try packaging anything else.  If we ever do end
up with a need for other software from this project, this derivation
could be extended and renamed, with tar2ext4 becoming an alias.