mautrix-telegram: fix crash
The previous backport of `mautrix-telegram`[1] appears to be a mistake
of mine (sorry for that!), unfortunately v0.9 isn't forward-compatible
due to a database migration, so this change can't be reversed.

The primary issue was that `telethon` was too old, so I decided to
upgrade it to 1.21 which is what I had on my 21.05-based tracking
branch while using `mautrix-telegram` anyways, so that's proven to work.

Closes #135583

[1] e1042038ac15ad7cf447fa1ed23a3746ee5d8331 / cd84b8ad8b2f64f6ccf1f27b910bcdd274585fb0