flexoptix-app: 5.9.0 -> 5.11.0 and fixes
This update is required because the autoupdater prevents the app from
running without being updated (which is not possible because it is a
store path).
So this updates the app and patches out the updater (essentially it
never thinks it needs to update, even though it prints on the command
line that it should).

Also fix the desktop item.

(cherry picked from commit 9aa05d18b77b20ecea5046fac4987a5130d9dfef)