haskell/configuration-hackage2nix: no x86_64-darwin → no aarch64-darwin
As a preparation for adding aarch64-darwin support, make sure that
whenever x86_64-darwin is unsupported, we also don't support
aarch64-darwin to avoid a mess in the diff and likely a lot of avoidable
build failures.

If it turns out to be supported we can always remove them later.