datadog-agent: 6.11.2 -> 7.30.2
datadog-agent is now built as a go module.  However, the build process
is now to be driven by the python invoke tool.  While datadog-agent
builds with buildGoModule, some investigation should be done to ensure
we are not missing any integrations or features from the invoke build
system.

This also updates datadog-agent from python 2.7 to 3.x.

One other issue of note: most of the invoke tasks seem to use some git
parsing to get version numbers and the like. This package may need to
be checked for reproducibility issues now.