xonotic: support user statistics reporting
Xonotic supports (opt-in) tracking of user statistics. For staistics to
be trackable each player needs a unique identifier that is based on some
cryptographic identity. For that to work we need to build the "blind"
library and provide it do xonotic during runtime.

The official xonotic releases ship with the public key of some sort of
central trust authority and thus we must retain that file within our
data package as well.

Before this commit the crypto_* commands in the ingame console weren't
available. With this commit you should be able to execute "crypto_keys"
commands as show below.

  $ crypto_keys
  0: public key key_0.d0pk (fingerprint: Xon//KssdlzGkFKdnnN4sgg8H+koTbBn5JTi37BAW1Q=)
     private ID key_0.d0si (public key fingerprint: 9piqrk8ajAPRc3BnSbkac5GA+yL3dOwh53BhyrfmVlI=)

Once the above command works you have to configure your player profile
to allow tracking (if you want) via the Multiplayer -> Profile dialog.
In the top right-hand corner there will be checkboxes to control the
level of tracking that you are opting in to.