gradle: Add ncurses dependencies on Linux
If ncurses is available Gradle can use it to prettify its output and
show progress during task execution. The fixup step patches the
bundled native libraries with libstdc++ but not with ncurses, which
means that Gradle will always use the (non-pretty) fallback native
platform libraries.

Patching the native platform libraries with ncurses5 and ncurses6
dependencies fixes the issue.

For some reason both ncurses5 and ncurses6 are required. Gradle will
refuse to start if, for example, the ncurses5 native platform jar is
missing.