nixos/etc: Force the source path to a string with context
With this change we ensure that the source path is always within the nix
store if passed as a `path` type. This used to work until refactored in
eb7120d where the path was converted to a string (implicitly).

This caused previously working configurations that propagated files in
/etc/ with files from the local (build time) filesystem to break. They
did now contain the absolute path to the files on the build host.

This surfaced in the initrd-networking-sshd test case where the ssh key
wasn't deployed onto the hosts again. If we want to  intentionally
change this behavior we should provide an alternative, proper warnings
and a changelog entry.