xmrig: etc
By default xmrig looks for a config file in the nix-store. This change makes it possible to override configFile argument which will write the `config.json` file.

I am still new to nix and nixpkgs and I do not know if this is the desired way to make this work, please let me know if should be done in some other way.