llvm_12: fix cross-compilation
Cross-compilation is broken because the method of finding ncurses has
changed, causing the build for the 'build system' to fail with a linking
error due to ncurses being for the 'host system' (where you're compiling
for).

This patch disables ncurses, which is not a very neat solution, but will
do until someone takes this upstream and gets it fixed properly.

Closes https://github.com/NixOS/nixpkgs/issues/127946.

Error that's seen before applying this:

/nix/store/hash-binutils-2.35.1/bin/ld: /nix/store/hash-ncurses-6.2-aarch64-unknown-linux-gnu/lib/libtinfo.so: error adding symbols: file in wrong format