acpica-tools: fix cross
The default value of INSTALLFLAGS is "-m 555 -s", -s being the option
to run the "strip" program on the installed files.  When
cross-compiling, we don't have a strip program (we have
"${stdenv.cc.targetPrefix}strip"), so install fails.

The simplest fix for this is to just remove -s from INSTALLFLAGS,
since stdenv will automatically strip all installed binaries at the
end anyway.