mosh: fix on aarch64-darwin
The build failure here is not due to Apple Silicon per se but instead
an artifact of the aarch64-darwin toolchain using a newer version of the
Xcode SDK as a basis. This causes issues building on case-insensitive
filesystems due to the collision between the standard <version> header
and a mosh source file called VERSION.

Fix with a MacPorts backport of an upstream mosh commit, via Homebrew.