git: ensure we are using the correct shell binary for tests
Previosuly the test scripts used /bin/sh which is a bit of an impurity.
It is mostly well-behaved but it essentially leaks the hosts state into
the build as /bin/sh points to some minimal shell implementation
configured on the host OS.

By patching the shebgangs of all the test scripts in the test folder
(t/*.sh) we can make sure that those run with the correct shell binary.

This was found as part of a random build failure of gitMinimal in
response to the systemd v249 PR being merged [0]. Since we have to
somehow touch the hash of the derivation to make the build failure go
away we might as well fix the hardcoded /bin/sh issue.

[0] https://github.com/NixOS/nixpkgs/commit/64556974b6674338a227d559975c9b4710f6e751#commitcomment-56385360