git: disable flaky test 6421-merge-partial-clone
Occasionally the test fails for unknown reasons but works as soon as we
change the derivation in any way. For now it is better to not have a
test that is flaky than having a test that occasionally breaks and
continously wastes time on debugging it.

This was found as part of a random build failure of gitMinimal in
response to the systemd v249 PR being merged [0].

[0] https://github.com/NixOS/nixpkgs/commit/64556974b6674338a227d559975c9b4710f6e751#commitcomment-56385360