catch: fix build w/glibc-2.34
The issue was that `SIGSTKSZ` isn't an actual const anymore and thus
can't be used to define sizes of static variables such as

    static char foo[SIGSTKSZ];

since this results in compiler errors such as

    error: array bound is not an integer constant before ']' token

Fedora worked around this by hard-coding the value in `catch`. Since
this is mainly a testing-framework and there's no other fix for v1 - we
should eventually remove it entirely in favor of v2 anyways - I guess
this is a good-enough fix.

Failing Hydra build: https://hydra.nixos.org/build/152455108