tpm2-tss: fix cross-compilation, enable strictDeps
It appears some dependencies were listed in checkInputs, they're moved
to the proper place in buildInputs & nativeBuildInputs.

Since this library was added as dependency of systemd, it's needed for
lots of packages to get cross-compilation working properly.

Also fixes pkgsStatic build, which fails due to cmocka not building.
Since we don't need it anyway when tests are disabled, gate it behind a
test for stdenv.buildPlatform == stdenv.hostPlatform.