maintainers/haskell: script for frozen package set (re)generation
Adds freeze-packages.sh which writes the frozen-packages.yaml config
file for hackage2nix. This is used for release branches to allow
continuous updates of the all-cabal-hashes package to get up to date
latest versions of packages in haskellPackages without updating the
default versions in that set (to avoid introducing breaking changes).

This is very useful for backports, since changing a version number in
the frozen-packages.yaml file and running 1-3 scripts is all that needs
to be done for that.