ghc on musl: Crude hack to allow integer-simple
as per discussion on
https://github.com/NixOS/nixpkgs/pull/130441#issuecomment-922275844 this
seems to be enough to make
```
nix-build -A pkgsMusl.haskell.packages.integer-simple.ghc884.hello
```
work again.

This is likely not the right patch (maybe `gmp` should be
`nativeBuildInput`? Or a propagated build input from the boot ghc?), but
this way I we can discuss (and I can pull the patch into my project...)