dhallPackages: Remove `dhall-packages`
Fixes https://github.com/NixOS/nixpkgs/issues/138374

The build failure in the above issue is due to
`pkgs.dhallPackages.dhall-packages` transitively depending on an
older Prelude version (12.0.0) that is incompatible with newer
Dhall versions.

I first attempted to upgrade `pkgs.dhallPackages.dhall-packages`, but
that failed due to a bad hash on an internal import, so I'm removing
the build instead.