pkgsStatic.perl: fix build
I don't understand how this worked prior to 1dbb306
("pkgsStatic.perl: Use `dontAddStaticConfigureFlags` instead of overriding"),
but what I do know is that it's very broken now, to the point that
perl-cross needs a patch.

I doubt the -Duseshrplib is actually necessary when native compiling,
but removing it would be a mass rebuild, so I've left it in place for
now.