mandoc: fix UTF-8-support detection and make more robust at runtime
locale(1) is not available in pkgsMusl.stdenv, but it is also not really
necessary. We just need to tell mandoc about *any* UTF-8 locale that is
also available at runtime.

For macOS C.UTF-8 is not available sadly, so we need to use
en_US.UTF-8. Using locale(1) for this is out of the question as NetBSD's
locale(1) depends on mandoc.