Revert "Add packageTests"
This reverts commit 87b3740d66b369e4e447f3fcb58a657a268f213d.

packageTests wasn't inherently wrong, but I am not fully convinced
that it is necessary. Specifically, it would only be used by:

 - `nix-build`, which is going to be a legacy command in the mid term
 - ofborg, perhaps, not even implemented

whereas the following use cases do not need it:

 - `nix build` (experimental command with a future, unlike `nix-build`)
 - `release.nix` (Hydra), might use similar logic but not reuse this

Further, considering that this is only necessary for tests of
functions, I am hesitant to introduce it.
Meanwhile, it should not hinder the rest of the PR.