quilt: wrap all required inputs
Some quilt commands fail on darwin with this error:

    stat: illegal option -- c
    usage: stat [-FlLnqrsx] [-f format] [-t timefmt] [file ...]

This is because at build time quilt is configured to use the coreutils
version of 'stat', but the 'quilt' wrapper does not wrap this coreutils
dependency and then at runtime the version from macOS is used (as found
in $PATH).

Fix by ensuring that coreutils is wrapped.

Also add and wrap the missing inputs for the other programs used by the
quilt commands: gawk, gnugrep, gnused, unixtools.column and
unixtools.getopt.