linux_hardened: make hardened patchset less picky
The hardened patches include the root Makefile, so they can add the
-hardened1 string to the end of the version. Unfortunately, they include
a context string which causes them to always fail to match if the minor
version is bumped.

In some cases, it's possible to fuzzy-apply the old patchset against a
newer minor version of the kernel. To accommodate this, we use
patchutils (as from e.g. fetchpatch) to strip all the context lines
from the root Makefile's patch only. This is probably a little more
aggressive than necessary: we only really want to do this for the first
few lines in the Makefile, but this is easiest and there aren't usually
any other changes to the file.

We could also just drop the patch to the Makefile, but this would remove
the -hardened string entirely from the kernel build, which probably
isn't desirable.