netbsd.compat: don't use objcopy on Darwin
objcopy on Darwin is a tale of woe and misery. As far as I can tell,
the NetBSD toolchain wants to use it to strip a bunch of sections from
the files it's outputting, but this doesn't work properly, at least on
aarch64-darwin.

The GCC version of objcopy appears to know how to read and write Mach-O
files, but reads in the Mach-O 64-bit file and writes it in a Mach-O 32-bit
container, which later causes ld to barf when it tries to link against it.

The LLVM version of objcopy just explodes when it finds some unimplemented
load commands.

In any case, just dropping the `objcopy -x` invocation _seems_ harmless enough,
and gets netbsd.xinstall working, which should fix a bunch of Python packages.