firefox-bin: Don't reference gnome in expression
Infix matching is used a heuristic in nixpkgs-update for what _not_ to
update which causes the firefox-bin packages not to get picked up by
the automatic updater.

While this "fix" isn't ideal and complicates the call-sites a bit I
think the trade-off is worth it to have up-to-date Firefox packages.